Move route code to dedicated folder
